Move Dimensions to its own module and fix tests
