+++ fixtures/8x8board.json
@@ -1,0 +1,1 @@
+{"game":{"id":"506514ef-249f-48b8-827b-7bf8d17ac7ad","ruleset":{"name":"royale","version":"v1.0.20"},"timeout":600},"turn":1,"board":{"height":8,"width":8,"snakes":[{"id":"gs_YkwKKSmYwqFFgDk9BycMvWf8","name":"PepperLongStocking🧦","latency":"370","health":99,"body":[{"x":0,"y":5},{"x":1,"y":5},{"x":1,"y":5}],"head":{"x":0,"y":5},"length":3,"shout":"0: 100 - 3"},{"id":"gs_vbvwfwk6jBc4jmCrKCbdJh3G","name":"does this work lol (unstable)","latency":"204","health":99,"body":[{"x":5,"y":6},{"x":5,"y":6},{"x":5,"y":6}],"head":{"x":5,"y":6},"length":3,"shout":""},{"id":"gs_6QpMpVPy7RpRxvcC9cc9V3xF","name":"Gaius Imperattlesnake","latency":"269","health":99,"body":[{"x":4,"y":1},{"x":5,"y":1},{"x":5,"y":1}],"head":{"x":4,"y":1},"length":3,"shout":""},{"id":"gs_6kQVWJXt9BFpD6dchrmX8qpM","name":"Nessegrev-beta","latency":"454","health":99,"body":[{"x":9,"y":0},{"x":9,"y":1},{"x":9,"y":1}],"head":{"x":9,"y":0},"length":3,"shout":""}],"food":[{"x":0,"y":4},{"x":4,"y":6},{"x":4,"y":0},{"x":8,"y":0},{"x":5,"y":5}],"hazards":[]},"you":{"id":"gs_vbvwfwk6jBc4jmCrKCbdJh3G","name":"does this work lol (unstable)","latency":"204","health":99,"body":[{"x":5,"y":6},{"x":5,"y":6},{"x":5,"y":6}],"head":{"x":5,"y":6},"length":3,"shout":""}}
